fix: do not copy timers and tickers in the default implementation
`DefaultClock.NewTimer()` makes a copy of a `time.Timer` instance
and stores it in the newly created `defaultTimer` instance:
```
func (dc DefaultClock) NewTimer(d time.Duration) Timer {
    return &defaultTimer{*time.NewTimer(d)}
}
```

Internally, `time.NewTimer()` adds the timer instance to the current `P`'s
timers heap and sets its state to `timerWaiting`.
After that, when `defaultTimer.Stop()` is called, the corresponding code
in `time.Timer.Stop()` sets the timer status to `timerDeleted` for
the *copied* instance, but the original instance in `P`'s heap is left intact.
As a result, the timer instance is not deleted from `P`'s heap.

Instead of copying, store the original pointer to `time.Timer` in
a `defaultTimer` instance. Same for `defaultTicker`.

Signed-off-by: Igor Bazhitov <igor.bazhitov@gmail.com>